refactor(milestones): split into a module
