--- contracts/escrow/src/refund.rs
@@ -1,2 +1,0 @@
-// Refund entrypoints are implemented in `contracts/escrow/src/lib.rs`.
-// This module retains refund-related helpers only.
