+++ contracts/escrow/src/types.rs
@@ -5,14 +5,7 @@ use soroban_sdk::{contracterror, contracttype, Address, String, Vec};
 #[allow(dead_code)]
 pub const CONTRACT_SUMMARY_SCHEMA_VERSION: u32 = 1;
 
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MilestoneSummary {
-    pub index: u32,
-    pub amount: i128,
-    pub released: bool,
-    pub refunded: bool,
-}
+use crate::milestones::{MilestoneSummary, ReleaseAuthorization};
 
 #[contracttype]
 #[derive(Clone, Debug, Eq, PartialEq)]
@@ -223,46 +216,6 @@ pub struct Contract {
     pub refunded_amount: i128,
     pub release_authorization: ReleaseAuthorization,
     pub reputation_issued: bool,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct Milestone {
-    pub amount: i128,
-    pub funded_amount: i128,
-    pub released: bool,
-    pub refunded: bool,
-    pub work_evidence: Option<String>,
-    pub refunded_amount: i128,
-    /// Optional Unix timestamp (seconds) after which the client may claim
-    /// a timeout refund for this milestone without arbiter involvement.
-    /// None means no deadline — the milestone never expires.
-    pub deadline: Option<u64>,
-}
-
-/// Defines who can approve milestone releases.
-#[contracttype]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum ReleaseAuthorization {
-    /// Only client can approve.
-    ClientOnly = 0,
-    /// Either client or arbiter can approve.
-    ClientAndArbiter = 1,
-    /// Only arbiter can approve.
-    ArbiterOnly = 2,
-    /// Both client and freelancer must approve; only either of them may release
-    /// after both approvals are present.
-    MultiSig = 3,
-}
-
-/// Tracks approval status for a milestone.
-/// Stored in temporary storage with TTL for expiry grace period.
-#[contracttype]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct MilestoneApprovals {
-    pub client_approved: bool,
-    pub freelancer_approved: bool,
-    pub arbiter_approved: bool,
 }
 
 #[contracttype]
